requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
numpy
yfinance
flask
plotly
pandas
openpyxl
azure-cosmos
azure-keyvault-secrets
azure-identity
fyers_apiv3
requests
opencensus
opencensus-ext-azure
redis
azure-storage-blob
azure-communication-email
azure-servicebus
# TA-Lib